+++ cmd/hidden_lake/helpers/adapters/common/values.go
@@ -1,5 +1,0 @@
-package common
-
-const (
-	HostService = "http://localhost"
-)