+++ CHANGELOG.md
@@ -6,10 +6,15 @@
 
 *??? ??, ????*
 
+### IMPROVEMENTS
+
+- Update `examples/echo_service`: append examples with docker-compose
+
 ### BUG FIXES
 
 - Update `cmd/hidden_lake/applications/filesharer`: fix creating hlf.stg
 - Update `cmd/hidden_lake/applications/filesharer|messenger`: fix tests with same ports
+- Update `cmd/hidden_lake/helpers/adapters/common`: fix print log
 
 <!-- ... -->